# services/api/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "exam-api",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v2/practice/sessions": {
      "post": {
        "tags": [
          "session-v2"
        ],
        "summary": "Create Session",
        "operationId": "create_session_api_v2_practice_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PracticeSessionCreateRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSessionEnvelopeV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/practice/sessions/{session_id}": {
      "get": {
        "tags": [
          "session-v2"
        ],
        "summary": "Get Session",
        "operationId": "get_session_api_v2_practice_sessions__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSessionEnvelopeV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/practice/sessions/{session_id}/answers": {
      "post": {
        "tags": [
          "session-v2"
        ],
        "summary": "Save Answers",
        "operationId": "save_answers_api_v2_practice_sessions__session_id__answers_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PracticeAnswerUpsertRequestV2"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAckV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/practice/sessions/{session_id}/submit": {
      "post": {
        "tags": [
          "session-v2"
        ],
        "summary": "Submit Session",
        "operationId": "submit_session_api_v2_practice_sessions__session_id__submit_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAckV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/practice/sessions/{session_id}/result": {
      "get": {
        "tags": [
          "session-v2"
        ],
        "summary": "Get Result",
        "operationId": "get_result_api_v2_practice_sessions__session_id__result_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSessionResultResponseV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/review/items/{item_id}": {
      "get": {
        "tags": [
          "review-v2"
        ],
        "summary": "Get Review Item",
        "operationId": "get_review_item_api_v2_review_items__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDetailResponseV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/review/items/{item_id}/redo": {
      "post": {
        "tags": [
          "review-v2"
        ],
        "summary": "Redo Review Item",
        "operationId": "redo_review_item_api_v2_review_items__item_id__redo_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAckV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/profile/overview": {
      "get": {
        "tags": [
          "profile-v2"
        ],
        "summary": "Get Profile Overview",
        "operationId": "get_profile_overview_api_v2_profile_overview_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOverviewResponseV2"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/profile/security": {
      "get": {
        "tags": [
          "profile-v2"
        ],
        "summary": "Get Profile Security",
        "operationId": "get_profile_security_api_v2_profile_security_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSecurityResponseV2"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "profile-v2"
        ],
        "summary": "Put Profile Security",
        "operationId": "put_profile_security_api_v2_profile_security_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileSecurityUpdateRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSecurityResponseV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/profile/goals": {
      "get": {
        "tags": [
          "profile-v2"
        ],
        "summary": "Get Profile Goals",
        "operationId": "get_profile_goals_api_v2_profile_goals_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileGoalsResponseV2"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "profile-v2"
        ],
        "summary": "Put Profile Goals",
        "operationId": "put_profile_goals_api_v2_profile_goals_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileGoalsUpdateRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileGoalsResponseV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/profile/info": {
      "get": {
        "tags": [
          "profile-v2"
        ],
        "summary": "Get Profile Info",
        "operationId": "get_profile_info_api_v2_profile_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileInfoResponseV2"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "profile-v2"
        ],
        "summary": "Put Profile Info",
        "operationId": "put_profile_info_api_v2_profile_info_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileInfoUpdateRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileInfoResponseV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperationAckV2": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "status"
        ],
        "title": "OperationAckV2"
      },
...
      "PracticeAnswerUpsertRequestV2": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/PracticeAnswerPayloadV2"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "type": "object",
        "title": "PracticeAnswerUpsertRequestV2"
      },
...
      "PracticeSessionCreateRequestV2": {
        "properties": {
          "track": {
            "type": "string",
            "enum": [
              "xingce",
              "essay"
            ],
            "title": "Track"
          },
          "entryKind": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Entrykind"
          },
          "paperCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Papercode"
          },
          "questionIds": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Questionids"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "track",
          "entryKind"
        ],
        "title": "PracticeSessionCreateRequestV2"
      },
      "PracticeSessionEnvelopeV2": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "track": {
            "type": "string",
            "title": "Track"
          },
          "entryKind": {
            "type": "string",
            "title": "Entrykind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PracticeSessionItemV2"
            },
            "type": "array",
            "title": "Items"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionLinkV2"
            },
            "type": "array",
            "title": "Actions"
          },
          "startedAt": {
            "type": "string",
            "title": "Startedat"
          },
          "submittedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submittedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "track",
          "entryKind",
          "status",
          "items",
          "actions",
          "startedAt"
        ],
        "title": "PracticeSessionEnvelopeV2"
      },
      "PracticeSessionItemV2": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "questionKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Questionkey"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "answerKind": {
            "type": "string",
            "title": "Answerkind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "prompt",
          "answerKind",
          "status"
        ],
        "title": "PracticeSessionItemV2"
      },
      "PracticeSessionResultResponseV2": {
        "properties": {
          "summary": {
            "items": {
              "$ref": "#/components/schemas/SummaryMetricV2"
            },
            "type": "array",
            "title": "Summary"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SectionCardV2"
            },
            "type": "array",
            "title": "Sections"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionLinkV2"
            },
            "type": "array",
            "title": "Actions"
          }
        },
        "type": "object",
        "required": [
          "summary",
          "sections",
          "actions"
        ],
        "title": "PracticeSessionResultResponseV2"
      },
      "ProfileGoalsResponseV2": {
        "properties": {
          "targetExam": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetexam"
          },
          "targetScore": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetscore"
          },
          "weeklyTargetHours": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weeklytargethours"
          }
        },
        "type": "object",
        "title": "ProfileGoalsResponseV2"
      },
      "ProfileGoalsUpdateRequestV2": {
        "properties": {
          "targetExam": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetexam"
          },
          "targetScore": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetscore"
          },
          "weeklyTargetHours": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weeklytargethours"
          }
        },
        "type": "object",
        "title": "ProfileGoalsUpdateRequestV2"
      },
      "ProfileInfoResponseV2": {
        "properties": {
          "displayName": {
            "type": "string",
            "title": "Displayname"
          },
          "realName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Realname"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          }
        },
        "type": "object",
        "required": [
          "displayName"
        ],
        "title": "ProfileInfoResponseV2"
      },
      "ProfileInfoUpdateRequestV2": {
        "properties": {
          "displayName": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Displayname"
          },
          "realName": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Realname"
          },
          "region": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          }
        },
        "type": "object",
        "required": [
          "displayName"
        ],
        "title": "ProfileInfoUpdateRequestV2"
      },
      "ProfileOverviewResponseV2": {
        "properties": {
          "summary": {
            "items": {
              "$ref": "#/components/schemas/SummaryMetricV2"
            },
            "type": "array",
            "title": "Summary"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SectionCardV2"
            },
            "type": "array",
            "title": "Sections"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionLinkV2"
            },
            "type": "array",
            "title": "Actions"
          }
        },
        "type": "object",
        "required": [
          "summary",
          "sections",
          "actions"
        ],
        "title": "ProfileOverviewResponseV2"
      },
      "ProfileSecurityResponseV2": {
        "properties": {
          "passwordSet": {
            "type": "boolean",
            "title": "Passwordset"
          },
          "emailBound": {
            "type": "boolean",
            "title": "Emailbound"
          },
          "phoneBound": {
            "type": "boolean",
            "title": "Phonebound"
          },
          "activeSessions": {
            "type": "integer",
            "title": "Activesessions"
          }
        },
        "type": "object",
        "required": [
          "passwordSet",
          "emailBound",
          "phoneBound",
          "activeSessions"
        ],
        "title": "ProfileSecurityResponseV2"
      },
      "ProfileSecurityUpdateRequestV2": {
        "properties": {
          "currentPassword": {
            "type": "string",
            "maxLength": 255,
            "minLength": 6,
            "title": "Currentpassword"
          },
          "newPassword": {
            "type": "string",
            "maxLength": 255,
            "minLength": 6,
            "title": "Newpassword"
          }
        },
        "type": "object",
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "title": "ProfileSecurityUpdateRequestV2"
      },
...
      "ReviewDetailResponseV2": {
        "properties": {
          "item": {
            "$ref": "#/components/schemas/ReviewItemV2"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ReviewAttemptOutV2"
            },
            "type": "array",
            "title": "History"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionLinkV2"
            },
            "type": "array",
            "title": "Actions"
          }
        },
        "type": "object",
        "required": [
          "item",
          "history",
          "actions"
        ],
        "title": "ReviewDetailResponseV2"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}